requirements.txt
flask
flask-socketio
gevent
gevent-websocket
gunicorn
qrcode
python-dotenv